backend/requirements.txt
fastapi
uvicorn[standard]
httpx
supabase
python-dotenv
apscheduler
